refactor(package): move filter classes into utils/filter
