+++ app/src/main/java/com/example/ezvault/data/FilterRepository.java
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
 import com.example.ezvault.viewmodel.SortedItemListView;
-import com.example.ezvault.utils.MainItemFilter;
+import com.example.ezvault.utils.filter.MainItemFilter;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
