+++ app/src/main/java/com/example/ezvault/utils/filter/ItemListFilter.java
@@ -1,10 +1,9 @@
-package com.example.ezvault.utils;
+package com.example.ezvault.utils.filter;
 
 import androidx.annotation.NonNull;
 
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.utils.ItemListView;
-import com.example.ezvault.utils.IItemFilter;
 
 import java.util.ArrayList;
 import java.util.Iterator;
